requirements.txt
protobuf
bitstring
pyserial
gprof2dot